Move and fix tcp and command adapter tests
